Review of pmic_io.{c,h}
Renamed functions:
- Pmic_commIntf_sendByte -> Pmic_ioTxByte
- Pmic_commIntf_recvByte -> Pmic_ioRxByte
- Pmic_calcCRC8 -> IO_calcCRC8 (and made private)

Other changes:
- Simplified implementations of Pmic_io{Tx,Rx}Byte
- Added documentation to public functions

+++ include/pmic.h
@@ -186,9 +186,9 @@ extern "C" {
  *
  * @{
  */
-#define PMIC_MAIN_INST (1U << 0U)
-#define PMIC_QA_INST   (1U << 1U)
-#define PMIC_NVM_INST  (1U << 2U)
+#define PMIC_MAIN_INST  (uint32_t)(1U << 0U)
+#define PMIC_QA_INST    (uint32_t)(1U << 1U)
+#define PMIC_NVM_INST   (uint32_t)(1U << 2U)
 /** @} */
 
 /**
